refactor: reorganize TIFF metadata tests and remove deprecated mode tests
